eval_framework/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "persona-eval"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "litellm>=1.0",
    "sentence-transformers>=2.7",
    "psycopg2-binary>=2.9",
    "click>=8.1",
    "slack-sdk>=3.27",
    "scipy>=1.13",
    "scikit-learn>=1.5",
    "numpy>=1.26",
    "textblob>=0.18",
    "vaderSentiment>=3.3",
    "httpx>=0.27",
]

[project.scripts]
evals = "persona_eval.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-timeout>=2.2",
    "hypothesis>=6.100",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "llm: tests requiring live LLM API calls",
    "slow: slow-running tests (>30s)",
    "gpu: tests requiring GPU acceleration",
]
